requirements.txt
ib_insync
pandas
numpy
python-dotenv
PyYAML
pytest
yfinance
scipy
